refactor: KMP-port ink.kt.mica — remove all JVM dependencies
Replace all JVM-specific types in 10 mica files for KMP compatibility:
- BigDecimal → Double (7 files: Expression, Story, StoryText, Declaration, Choice, ConditionalOption, StoryLoader)
- java.io.InputStream → String input (InkParser, StoryWrapper)
- Jackson JsonParser/JsonGenerator → kotlinx.serialization.json (StoryLoader, StorySaver)
- Java reflection → GameObjectResolver fun interface (Expression)
- MathContext/RoundingMode → kotlin.math.pow (Expression)

Update cross-implementation gap analysis in ink.kt.puml and ink.java.puml:
- Add mica KMP porting status table (6 JVM deps → KMP replacements)
- Update mica column: BigDecimal → Double in progress tracker
- Add Java→ink.kt callback gap table (5 callbacks Java lacks)
- Update porting priority: mica KMP port marked ✅ Done

+++ ink-kmp-mcp/src/commonMain/kotlin/ink/kt/mica/Choice.kt
@@ -1,7 +1,6 @@
 package ink.kt.mica
 
 import ink.kt.mica.util.InkRunTimeException
-import java.math.BigDecimal
 
 internal class Choice(
     textBase: String,
@@ -57,7 +56,7 @@ internal class Choice(
             try {
                 val obj = Declaration.evaluate(condition, story)
                 if (obj is Boolean && !obj) return false
-                if (obj is BigDecimal && obj.toInt() <= 0) return false
+                if (obj is Double && obj.toInt() <= 0) return false
             } catch (e: InkRunTimeException) {
                 story.logException(e)
                 return false
